IPCThreadState: Isolate parcels in transact() and support nested binder calls
Old legacy binderized drivers might call severval public methods directly
when currently being handled by IPCThreadState itself. This caused corruption
in Parcels mOut and mIn and triggers unsupported Parcel.Remove() killing the
process.

By isolating mIn and mOut Parcels and detect if we are already talking to the
driver, we can postpone the nested binder calls of the driver and merge these
in the next interation.

Change-Id: I0bd0389de3078b1c1640e8000282fe26d2b5541f